+++ eoreader/products/optical/s3_product.py
@@ -943,9 +943,7 @@ class S3Product(OpticalProduct):
         """
         try:
             if self.is_archived:
-                quicklook_path = self.path / path.get_archived_path(
-                    self.path, file_regex=r".*.jpg"
-                )
+                quicklook_path = self._get_archived_path(file_regex=r".*.jpg")
             else:
                 quicklook_path = str(next(self.path.glob("**/*.jpg")))
         except (FileNotFoundError, StopIteration):